// backend/src/modules/auth/auth.schemas.ts
import { z } from "zod";

export const registerSchema = z.object({
  email: z.string().email(),
  password: z.string().min(8),
  name: z.string().min(1).max(80).optional(),
  preferredLanguage: z.enum(["mn", "en"]).default("mn")
});

export const loginSchema = z.object({
  email: z.string().email(),
  password: z.string().min(8)
});